makes intervalType optional
Because it was always being set to a placeholder, that placeholder string was being used even when intervalType was not passed into the function. And intervalType is not relevant for a non-jetpack site. So this broke the "back to your site" link in the "thank-you" page after purchasing a plan.

I don't think we should be using these default values at all, but I'm not positive where else this might be used, so I'd rather just get the bug fixed right away. I think we should just remove this whole abstraction layer for routing in plans. It's adding more complexity and confusion, and it's not consistent with how we do routing elsewhere in calypso.